requirements.txt
python-dotenv
sse_starlette
sentence-transformers
uvicorn
dashscope
pymupdf
langserve
langchain-community
langchain_openai
langchain-core
langchain_chroma
langgraph
langchain-milvus
chromadb
xinference_client
pyjwt
zhipuai
elasticsearch
jieba
nltk
mysql-connector-python
PyMySQL